requirements.txt
Django
torch
numpy
Pillow
opencv-python
ultralytics
